+++ parser.d.ts
@@ -12,7 +12,8 @@ type GetLastTag<I> = I extends `${string}${Combinators}${infer Right}`
     : GetLastTag<Right>
   : I
 
-type PseudoClassesFirstChar =
+type PseudoFirstChar =
+  | ':'
   | 'a'
   | 'b'
   | 'c'
@@ -108,9 +109,9 @@ type PostprocessEachUnchecked<I> =
     ? PostprocessEachUnchecked<`${Tag}&${Rest}`>
     : I extends `${infer Tag}#${string}`
     ? PostprocessEachUnchecked<Tag>
-    : I extends `${infer Tag}:${PseudoClassesFirstChar}${string}&${infer Rest}`
+    : I extends `${infer Tag}:${PseudoFirstChar}${string}&${infer Rest}`
     ? PostprocessEachUnchecked<`${Tag}&${Rest}`>
-    : I extends `${infer Tag}:${PseudoClassesFirstChar}${string}`
+    : I extends `${infer Tag}:${PseudoFirstChar}${string}`
     ? PostprocessEachUnchecked<Tag>
     : I extends `${string}|${infer Tag}` // namespace prefix
     ? PostprocessEachUnchecked<Tag>