+++ parser.test.ts
@@ -357,6 +357,7 @@ type _Tests = [
   >,
   Expect<Equal<ParseSelector<':func(arg)'>, Element>>,
   Expect<Equal<ParseSelector<'::func(arg)'>, Element>>,
+  Expect<Equal<ParseSelector<'a::-p-text(Hello)'>, HTMLAnchorElement>>,
 ]
 
 declare function delegate<